// src/enums/ListName.ts
export enum ListName {
    Details = 'Detaily',
    Requests = 'Zadosti',
    QA = 'QA',
    Categories = 'Kategorie',
    Duties = 'Povinnosti',
    Forms = 'Formulare',
    RequiredDocuments = 'PozadovaneDokumenty',
    DocumentTypes = 'TypyDokumentu',
    RequestTypes = 'TypyZadosti',
    Notifications = 'Notification',
    WorkflowActions = 'WorkflowActions',
    ActionToTrigger = 'ActionToTrigger',
    Logs = 'Logs'
}